Refactor: Move quote verification tests to dedicated files
This commit reorganizes the test suite by moving TDX and SGX quote verification tests into their own files. It also introduces a new `qvl-helpers.ts` file to share common testing utilities and constants. This improves test organization and maintainability.
